Clean runtime address checks and build workflow

+++ mem.cpp
@@ -3,14 +3,14 @@
 
 
 
-void mem::PatchEx(BYTE* dst, BYTE* src, unsigned int size, HANDLE hProcess)
+void mem::PatchEx(BYTE* dst, BYTE* src, size_t size, HANDLE hProcess)
 {
 	DWORD oldprotect;
 	VirtualProtectEx(hProcess, dst, size, PAGE_EXECUTE_READWRITE, &oldprotect);
 	WriteProcessMemory(hProcess, dst, src, size, nullptr);
 	VirtualProtectEx(hProcess, dst, size, oldprotect, &oldprotect);
 }
-void mem::NopEx(BYTE* dst, unsigned int size, HANDLE hProcess)
+void mem::NopEx(BYTE* dst, size_t size, HANDLE hProcess)
 {
 	BYTE* nopArray = new BYTE[size];
 	memset(nopArray, 0x90, size);
